more updates
yukari is working great now
I think there's some problems with eirin's spell, not sure.
updated a number of ai quirks and bugs too.
question is, do we want a skin for everybody right now? or I will just leave it and say go?

also, I think we are all done except the foreword, so, you know, get on that?