some more bug fixes in general
changed some prompts and messages
changed the structure of shan, so it can be properly interacted with
also remilia bug fix
alice seems to have a bug too, but I can't quite see what the problem is.
fixed a bug for kurumi, but I am not sure if the patch works.