updates to chess mode
chess mode being the biggest update.
I rolled the treasure system into the game. unknown if there's bugs, so that requires more testing.
especially that treasures can spawn on top of players somehow.

I think I will leave akyuu as is for now. It's not like she can be unlocked easily anyway, and not having a third skill doesn't detract too much from the cool factor. so I have some time to patch something in if I need to.

also fixed some bugs with letty, patchy and flan.

also 2B.
Now that I wrote it, like, 2B is really simplistic. Not that it's bad, because it is true that Nier's combat is pretty bare bones, and she's just battle type anyway, but I do wish there's a little more to her.

That being said, being simple makes her a bit boring but not unfun. So I guess it's cool.