bug finally fixed
so
removeequiptrigger only removes the triggers from the equip skills.
the equip skills with enable can still be used, not because enable was not removed, but because getSkills function is used to check for skills at the enable timings (i.e. choose to use). getSkills only checks for which equip card you have, and does not check if they are actually trigger removed or not.
in other words, addequiptrigger and removeequiptrigger only affects equip skills with trigger. equip skills without trigger are managed using getSkills (which renews state at every instance of enable triggers, and allows it to naturally reflect the gaining and loss of equips without need for specific update functions).
to combat this problem, a new classlist tag, 'unequip' was introduced, to indicate a card has its skills removed and should not count towards getSkills.
note that adding 'removing' to the card's classlist can cause a similar effect, but have a large amount of side effects (mostly involving treating the card as complete non-existent) that it should not be used at all.

also, fixed the brawl constant bug.

I swear to god, why is there so many hardcodes in this fucking thing.
although it is kinda my fault for not reading the whole thing...