Ok
so, I am up to adding an accessible rule button for every mode.
just need to take the setpopped and place it in the gamestart of every mode, and we should be golden.
oh, and adjust its size.

While I am here, I am gonna rant.
I realize that absolutely nobody is reading this and even if someone is, they would have no clue what I am ranting about.

well, I don't care.

The original designs of this goddamn game is so fucking obnoxious.
The UI looks nice. The frames are good looking. The actions are not too smooth but acceptable. HTML and CSS allows much better control and much more design elements than C++. Not needing to write a fucking QT translation patch is absolutely beautiful.
However, that's about where the praise ends. The system design is fine, but just, so fucking much of the UI design decisions is horrible.
Hey, we have card text. We are going to fit the card text into the world's smallest readable fucking box so that no matter how short a text is, it would always look like a goddamn spagetti. seriously, the box can hold like 8 words on one line. EIGHT. a single sentence is like 15 to 30 words, and that's just the short ones. once we get to 150...oh boy. That's not even a long paragraph! But its going to be 20 lines of text and look like a goddamn wall and you are gonna like it.
Oh and it scrolls at the like 7 lines so have fun with that. Not to mention the ridicious placement of the skill name itself.
Hey, here's a grid, which have 12 columns. We are going to fit words into them. 12 words a row is not actually that bad, if you don't take 4 columns out and just leave them empty for no goddamn reason!
I swear to god, if people realize how easy it is to change CSS files, nobody will be putting up with this shit.

the problem with text doesn't even end there. The game is actually rather helpful. There are tooltips in the settings that you can read. It would be a nice addition, if you didn't have to click four times into THE MOST HIDDEN-AWAY MENU IN THE ENTIRE GAME to get even read them!
They are literally in the last menu of the last menu, and is the only place in the game that has 3 layers of menu.
Oh and there's absolutely nothing indicating the help docs even exists, and the help docs shares the textbox with the other one I just ranted at.
The developer tunnels on developped vision so hard it's actually kind of fasinacting at this point.
It took me a year to even notice there was even a help doc in the game. and I develop this game every week. I only found about this thing because I was going through every string in a script for some reason and just came across it.
The tunnel vision only gets more insane too. There are 3 extra game modes the developper has designed and developed himself, into the game, along with several expansion packs. Which is fairly neat. They are somewhat interesting, usually not too overpowered, and is generally a fresh breath of life into the game.
But that is also a hugeass problem.
Because, as the gamer, I should not be forced to download 200 to 300 extra sets of assets to even run the game.
If you clearly know that what you made is not what the consumer is coming for, you can still roll the expansion in, but fuck it should not take double the amount of space and assets the original base game takes.
Its like, when somebody buys Catan, you package in 4 expansions, charge them extra (in terms of space consumption), and just sells it to them without even telling them what is in it.
And as the cherry on top, also package all the instruction manuals into the smallest baggie attached to the bottom of one of the resources cards in one of the expansions and pretend you didn't know it existed.
Needless to say, these things might as well as not exist to the general player.
But the consequence doesn't even end at the goddamn space consumption.

There is so much goddamn garbage code and useless settings in the game engine that I am actually kind of amazed.

Yes, this framework is rather powerful tool, but goddamn if it is not butchered by developer tunnel vision and simply stupid design.

Out of the 230 commits I have made, I think like a third of it is adding quality of life and UI changes. I can't even go back to the basegame now cuz its just so unbearable.

well, rant over.
hopefully nobody ever reads this.