right, i should be slower and actually write what is happening.
so I forgot to add the mobile class to the 三 button. fixed it.
turns out the massive repetitve lines of adding the same div, was actually to literally draw the 三 using CSS. There's a fine line between genius and stupidity, and I am not going to comment further than that.

turns out, if you failed to download any piece of asset during an asset download (regardless of reason) the version number actually does not update, and is possible to start an infinity loop. fixed.
now I am trying to make coding.net the default update link as opposed to github.