thinned down a lot of code
so I noticed there's a lot of redundent stuff here.

not just because some rules I have removed, but also because the designer was a douche who likes to put a function into the main game for ridiciolus (pretty sure i misspelt that) reasons.

I am not going to make a list of what I removed, because some of the stuff I don't even know what they do. I removed them because they are not used anywhere, or is entirely self-referential.
I also removed all mentions of 'stone', 'tafang', and 'jiange' modes. I also removed every mention of hujia, seeing as I think its dumb and will not be used in design, maybe ever.

There's for sure going to be bugs. So keep an eye out.