updates to characters and bgm
so, that was a crazy, like, 12 updates, which in total accomplished all of nothing?

well.
as it turns out, the tencent platform pages server, despite indicating that compliation is complete, is actually not, and actually takes about 1 hour for the website to be updated. I thought I was crazy for a few hours, but it turns out I just had to wait.
At least I cleaned up everything afterwards.

updates:
 sb incident now usable, and has a mark to indicate which characters have already used spell card.
rylai and aya has used the same format to check spell card activations.
so, global:useSkillAfter does not, in fact, recognize any skills activated on a global scale. even with the global tag. In fact, even with a global skill to check player:useSkillAfter does not work at all. It is possible it does not work due to the test condition, which has the main camera constantly rotated, and might mess up the global timing. More testing and code reading is required.
background music has now been streamlined.
background music now is updated to use on coding.me, by accessing a different place on the web instead of the current repository.
Also, playbackgroundmusic is updated: has 3 parameters: name, name of music to be played; temp: indicates if the current music is temporary (to update a music storage, currently unused), and marisa, which, if true, indicates the music does override marisa's theme.
every music change should know be written this way: game.playBackgroundMusic(name)
also, look for new bgm for saitama fight. silence is really boring.