some quick updates
shuffled setting around. tried to add section separators, unknown effect on phone.
fixed visual bug with frog. previously witnessed bug with frog, but was not able to replicate. more testing required.
a lot of weird stuff assoicated with "isMine()" and "auto" interactions.
fixed bug with homura, inelegant though.
problem with homura is more with choosetouse section.
choosetouse accomadates choosebutton skills, but use it weirdly.
choosetouse resets the process if choosebutton is detected, but choosebutton itself is also a choosetouse process.
such a process is:
choosetouse (result: choosebutton)
choosetouse (result: button.links)
choosetouse (result: true)
in other words, the action of finishing a choosebutton dialog is considered finishing the choosetouse process.
In fact, using any skills is considered finishing the choosetouse process, as "useSkill" is an accepted result. Instead, the only accepted result should be either the filtered card, or false (or endturn).
lots of complication with in and out of your phase, ai, ismine interactions.