puzzles!
puzzle omde is introduced!

also, bug with backgroundmusic due to directstart needs to be fixed still.

directstart, when used,, does not arena properly, and does not generate either a ui.window nor a music player.

trying to add a music player in the function does not work either.

however, the music does play if selected from the config menu.
In other words, the arena is previously generated, and is disconnected from the current functions perhaps, such as being another instance, and cannot do functions in the new instance, but the old instance can call its own functions.

the same bug applies for "no splash" setting, however, and it's unknown if it does indeed generate 2 instances for some reason.
which might also be why one cannot run any commands from the console.

should the hypothese be true, there would be no real way to fix such bug?